+++ rust/dbn/src/decode/dbn/sync.rs
@@ -84,6 +84,12 @@ where
     pub fn into_inner(self) -> R {
         self.decoder.into_inner()
     }
+
+    /// Sets the behavior for decoding DBN data of previous versions.
+    pub fn set_upgrade_policy(&mut self, upgrade_policy: VersionUpgradePolicy) {
+        self.metadata.upgrade(upgrade_policy);
+        self.decoder.set_upgrade_policy(upgrade_policy);
+    }
 }
 
 impl<'a, R> Decoder<zstd::stream::Decoder<'a, BufReader<R>>>
@@ -252,6 +258,11 @@ where
             self.version = version;
             Ok(())
         }
+    }
+
+    /// Sets the behavior for decoding DBN data of previous versions.
+    pub fn set_upgrade_policy(&mut self, upgrade_policy: VersionUpgradePolicy) {
+        self.upgrade_policy = upgrade_policy;
     }
 
     /// Returns a mutable reference to the inner reader.