+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.14.2 - 2023-11-17
+### Enhancements
+- Added `set_upgrade_policy` setters to `DbnDecoder`, `DbnRecordDecoder`,
+  `AsyncDbnDecoder`, and `AsyncDbnRecordDecoder`
+- Added `from_schema` classmethod for Python `RType` enum
+
+### Breaking changes
+- Renamed parameter for Python Enum classmethod constructors to `value` from `data`.
+
 ## 0.14.1 - 2023-11-17
 ### Enhancements
 - Added new trait `compat::SymbolMappingRec` for code reuse when working with